+++ formats/src/main/scala/overflowdb/formats/graphml/GraphMLExporter.scala
@@ -15,8 +15,6 @@ import scala.jdk.CollectionConverters.{IteratorHasAsScala, MapHasAsScala}
  * http://graphml.graphdrawing.org/primer/graphml-primer.html
  * */
 object GraphMLExporter extends Exporter {
-  val KeyForNodeLabel = "labelV"
-  val KeyForEdgeLabel = "labelE"
 
   override def runExport(graph: Graph, outputRootDirectory: Path) = {
     val outFile = resolveOutputFile(outputRootDirectory)
@@ -97,32 +95,4 @@ object GraphMLExporter extends Exporter {
       s"""<data key="$encodedPropertyName">$propertyValue</data>"""
     }.mkString("\n")
   }
-
-  object Type extends Enumeration {
-    val Boolean = Value("boolean")
-    val Int = Value("int")
-    val Long = Value("long")
-    val Float = Value("float")
-    val Double = Value("double")
-    val String = Value("string")
-
-    def fromRuntimeClass(clazz: Class[_]): Type.Value = {
-      if (clazz.isAssignableFrom(classOf[Boolean]) || clazz.isAssignableFrom(classOf[java.lang.Boolean]))
-        Type.Boolean
-      else if (clazz.isAssignableFrom(classOf[Int]) || clazz.isAssignableFrom(classOf[Integer]))
-        Type.Int
-      else if (clazz.isAssignableFrom(classOf[Long]) || clazz.isAssignableFrom(classOf[java.lang.Long]))
-        Type.Long
-      else if (clazz.isAssignableFrom(classOf[Float]) || clazz.isAssignableFrom(classOf[java.lang.Float]))
-        Type.Float
-      else if (clazz.isAssignableFrom(classOf[Double]) || clazz.isAssignableFrom(classOf[java.lang.Double]))
-        Type.Double
-      else if (clazz.isAssignableFrom(classOf[String]))
-        Type.String
-      else
-        throw new AssertionError(s"unsupported runtime class `$clazz` - only ${Type.values.mkString("|")} are supported...}")
-    }
-  }
-
-  private case class PropertyContext(name: String, tpe: Type.Value)
 }
