+++ formats/src/main/scala/overflowdb/formats/graphml/GraphMLImporter.scala
@@ -4,7 +4,7 @@ import overflowdb.Graph
 import overflowdb.formats.Importer
 
 import java.nio.file.Path
-import scala.xml.XML
+import scala.xml.{NodeSeq, XML}
 
 /**
  * Imports GraphML into OverflowDB
@@ -18,17 +18,39 @@ object GraphMLImporter extends Importer {
   override def runImport(graph: Graph, inputFiles: Seq[Path]): Unit = {
     assert(inputFiles.size == 1, s"input must be exactly one file, but got ${inputFiles.size}")
     val doc = XML.loadFile(inputFiles.head.toFile)
+
+    val keyEntries = doc \ "key"
     val graphXml = doc \ "graph"
 
-    for (node <- graphXml \ "node") {
-      addNode(graph, node)
+    { // nodes
+      val nodePropertyContextById = parsePropertyEntries("node", keyEntries)
+      for (node <- graphXml \ "node") {
+        addNode(graph, node, nodePropertyContextById)
+      }
     }
-    for (edge <- graphXml \ "edge") {
-      addEdge(graph, edge)
+
+    { // edges
+      val edgePropertyContextById = parsePropertyEntries("edge", keyEntries)
+      for (edge <- graphXml \ "edge") {
+        addEdge(graph, edge, edgePropertyContextById)
+      }
     }
   }
 
-  private def addNode(graph: Graph, node: scala.xml.Node): Unit = {
+  private def parsePropertyEntries(forElementType: String, keyEntries: NodeSeq): Map[String, PropertyContext] = {
+    keyEntries
+      .filter(_ \@ "for" == forElementType)
+      .map { node =>
+        val id = node \@ "id"
+        val name = node \@ "attr.name"
+        val graphmlType = node \@ "attr.type"
+        val tpe = Type.withName(graphmlType)
+        (id, PropertyContext(name, tpe))
+      }
+      .toMap
+  }
+
+  private def addNode(graph: Graph, node: scala.xml.Node, propertyContextById: Map[String, PropertyContext]): Unit = {
     val id = node \@ "id"
     var label: Option[String] = None
     val keyValuePairs = Seq.newBuilder[String]
@@ -36,7 +58,7 @@ object GraphMLImporter extends Importer {
     for (entry <- node \ "data") {
       val value = entry.text
       entry \@ "key" match {
-        case "labelV" => label = Option(value)
+        case KeyForNodeLabel => label = Option(value)
         case propertyName => keyValuePairs.addAll(Seq(propertyName, value))
       }
     }
@@ -47,7 +69,7 @@ object GraphMLImporter extends Importer {
     } graph.addNode(id, label, keyValuePairs.result: _*)
   }
 
-  private def addEdge(graph: Graph, edge: scala.xml.Node): Unit = {
+  private def addEdge(graph: Graph, edge: scala.xml.Node, propertyContextById: Map[String, PropertyContext]): Unit = {
     val sourceId = edge \@ "source"
     val targetId = edge \@ "target"
     var label: Option[String] = None
@@ -56,7 +78,7 @@ object GraphMLImporter extends Importer {
     for (entry <- edge \ "data") {
       val value = entry.text
       entry \@ "key" match {
-        case "labelE" => label = Option(value)
+        case KeyForEdgeLabel => label = Option(value)
         case propertyName => keyValuePairs.addAll(Seq(propertyName, value))
       }
     }
